refactor(update): extract update system into par-term-update subcrate
Create new par-term-update crate containing:
- http.rs: HTTP agent helper with native-tls + download_file utility
- manifest.rs: File tracking system for bundled assets (SHA-256 based)
- self_updater.rs: In-place binary replacement for standalone installs
- update_checker.rs: GitHub release polling with configurable frequency

Dependencies: par-term-config, ureq, native-tls, chrono, semver,
sha2, zip, serde, serde_json, log.

src/manifest.rs, src/self_updater.rs, src/update_checker.rs replaced
with thin `pub use par_term_update::<module>::*;` re-export shims.

Fixed internal references:
- crate::config::{Config, UpdateCheckFrequency} -> par_term_config::
- crate::shader_installer::download_file() -> crate::http::download_file()

Closes #179
