+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["par-term-acp", "par-term-config", "par-term-fonts", "par-term-keybindings", "par-term-render", "par-term-scripting", "par-term-settings-ui", "par-term-ssh", "par-term-terminal"]
+members = ["par-term-acp", "par-term-config", "par-term-fonts", "par-term-keybindings", "par-term-render", "par-term-scripting", "par-term-settings-ui", "par-term-ssh", "par-term-terminal", "par-term-update"]
 resolver = "2"
 
 [package]
@@ -52,6 +52,9 @@ par-term-scripting = { path = "par-term-scripting", version = "0.1.0" }
 
 # SSH host management and discovery
 par-term-ssh = { path = "par-term-ssh", version = "0.1.0" }
+
+# Self-update and update-check system
+par-term-update = { path = "par-term-update", version = "0.1.0" }
 
 # Terminal emulator core (with OSC 7/1337 hostname/cwd events, RemoteHost, and OSC 934 named progress bars)
 par-term-emu-core-rust = { version = "0.39.0", default-features = false, features = ["rust-only"] }
