+++ par-term-update/Cargo.toml
@@ -1,0 +1,44 @@
+[package]
+name = "par-term-update"
+version = "0.1.0"
+edition = "2024"
+authors = ["Paul Robello <probello@gmail.com>"]
+description = "Self-update and update-check system for par-term terminal emulator"
+license = "MIT"
+repository = "https://github.com/paulrobello/par-term"
+homepage = "https://github.com/paulrobello/par-term"
+keywords = ["terminal", "update", "self-update"]
+categories = ["command-line-utilities"]
+
+[dependencies]
+# Configuration types (Config, UpdateCheckFrequency)
+par-term-config = { path = "../par-term-config", version = "0.1.0" }
+
+# HTTP client for downloading updates and checking releases
+ureq = { version = "3.2.0", default-features = false, features = ["native-tls", "gzip"] }
+native-tls = "=0.2.16"
+
+# Date/time for tracking last check timestamp
+chrono = "0.4.43"
+
+# Semantic version comparison
+semver = "1.0.27"
+
+# SHA-256 hashing for binary verification
+sha2 = "0.10.9"
+
+# ZIP archive extraction for binary updates
+zip = "8.1.0"
+
+# Serialization for manifest tracking
+serde = { version = "1.0.228", features = ["derive"] }
+serde_json = "1.0.149"
+
+# Error handling
+anyhow = "1.0.101"
+
+# Logging
+log = "0.4.29"
+
+[dev-dependencies]
+tempfile = "3.25"
